refactor: update shared to ver
