+++ chk/modules/ver/constants.py
